+++ asyncapi-template/partials/SimDeviceSim.java
@@ -2,7 +2,7 @@
 // IT IS AUTOMATICALLY GENERATED BY THE 'generateDeviceFiles' GRADLE TASK
 // FROM THE TEMPLATE FILE 'asyncapi-template/partials/SimDeviceSim.java'.
 // EDIT THAT FILE INSTEAD.
-package org.team199.wpiws.devices;
+package org.carlmontrobotics.wpiws.devices;
 
 import java.math.BigDecimal;
 import java.util.Arrays;
@@ -17,12 +17,12 @@ import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.function.Predicate;
 
-import org.team199.wpiws.Pair;
-import org.team199.wpiws.StateDevice;
-import org.team199.wpiws.connection.ConnectionProcessor;
-import org.team199.wpiws.connection.WSValue;
-import org.team199.wpiws.interfaces.SimDeviceCallback;
-import org.team199.wpiws.interfaces.ObjectCallback;
+import org.carlmontrobotics.wpiws.Pair;
+import org.carlmontrobotics.wpiws.StateDevice;
+import org.carlmontrobotics.wpiws.connection.ConnectionProcessor;
+import org.carlmontrobotics.wpiws.connection.WSValue;
+import org.carlmontrobotics.wpiws.interfaces.SimDeviceCallback;
+import org.carlmontrobotics.wpiws.interfaces.ObjectCallback;
 
 /**
  * Represents a miscellaneous simulated device
@@ -252,7 +252,7 @@ public class SimDeviceSim extends StateDevice<SimDeviceSim.State> {
     }
 
     /**
-     * An implementation of {@link org.team199.wpiws.interfaces.DeviceMessageProcessor} which processes WPI HALSim messages for SimDeviceSims
+     * An implementation of {@link org.carlmontrobotics.wpiws.interfaces.DeviceMessageProcessor} which processes WPI HALSim messages for SimDeviceSims
      * @param device the device identifier of the device sending the message
      * @param data the data associated with the message
      */
