+++ asyncapi-template/template/$$schema$$.java
@@ -16,7 +16,7 @@
 // IT IS AUTOMATICALLY GENERATED BY THE 'generateDeviceFiles' GRADLE TASK
 // FROM THE TEMPLATE FILE 'asyncapi-template/template/$$schema$$.java'.
 // EDIT THAT FILE INSTEAD.
-package org.team199.wpiws.devices;
+package org.carlmontrobotics.wpiws.devices;
 
 import java.util.List;
 {% if hasId -%}
@@ -32,12 +32,12 @@ import java.util.concurrent.CopyOnWriteArraySet;
 import java.util.function.BiConsumer;
 import java.util.function.Consumer;
 
-import org.team199.wpiws.StateDevice;
-import org.team199.wpiws.connection.ConnectionProcessor;
-import org.team199.wpiws.connection.WSValue;
-import org.team199.wpiws.interfaces.*;
+import org.carlmontrobotics.wpiws.StateDevice;
+import org.carlmontrobotics.wpiws.connection.ConnectionProcessor;
+import org.carlmontrobotics.wpiws.connection.WSValue;
+import org.carlmontrobotics.wpiws.interfaces.*;
 {%- if (props | usesCustomTypes(name)) %}
-import org.team199.wpiws.types.*;
+import org.carlmontrobotics.wpiws.types.*;
 {%- endif %}
 
 /**
@@ -254,7 +254,7 @@ public class {{ name }}Sim {
     {% endfor -%}
 
     /**
-     * An implementation of {@link org.team199.wpiws.interfaces.DeviceMessageProcessor} which processes WPI HALSim messages for {{ name }}Sims
+     * An implementation of {@link org.carlmontrobotics.wpiws.interfaces.DeviceMessageProcessor} which processes WPI HALSim messages for {{ name }}Sims
      * @param device the device identifier of the device sending the message
      * @param type the type of the device sending the message
      * @param data the data associated with the message
