+++ src/main/java/org/carlmontrobotics/wpiws/connection/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Manages low level interfacing with WebSockets and basic WPI HALSim protocol processing and formatting
+ */
+package org.carlmontrobotics.wpiws.connection;
