+++ src/main/java/org/carlmontrobotics/wpiws/interfaces/LongCallback.java
@@ -1,0 +1,16 @@
+package org.carlmontrobotics.wpiws.interfaces;
+
+/**
+ * Represents a device callback which accepts a long value as data
+ */
+@FunctionalInterface
+public interface LongCallback {
+
+    /**
+     * Calls the callback function
+     * @param name the device identifier of the device calling the callback
+     * @param value the long data value
+     */
+    public void callback(String device, long value);
+
+}
