+++ src/main/java/org/carlmontrobotics/wpiws/interfaces/ObjectCallback.java
@@ -1,0 +1,16 @@
+package org.carlmontrobotics.wpiws.interfaces;
+
+/**
+ * Represents a device callback which accepts an object type as data
+ */
+@FunctionalInterface
+public interface ObjectCallback<T> {
+
+    /**
+     * Calls the callback function
+     * @param name the device identifier of the device calling the callback
+     * @param value the data value
+     */
+    public void callback(String name, T value);
+
+}
