+++ src/main/java/org/carlmontrobotics/wpiws/interfaces/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * All required {@link java.lang.FunctionalInterface} interfaces used in this API
+ */
+package org.carlmontrobotics.wpiws.interfaces;
