+++ src/main/java/org/carlmontrobotics/wpiws/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Code to interface with the WebSockets API of WPILib
+ */
+package org.carlmontrobotics.wpiws;
