--- src/main/java/org/team199/wpiws/connection/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Manages low level interfacing with WebSockets and basic WPI HALSim protocol processing and formatting
- */
-package org.team199.wpiws.connection;
