--- src/main/java/org/team199/wpiws/interfaces/DoubleCallback.java
@@ -1,16 +1,0 @@
-package org.team199.wpiws.interfaces;
-
-/**
- * Represents a device callback which accepts a double value as data
- */
-@FunctionalInterface
-public interface DoubleCallback {
-
-    /**
-     * Calls the callback function
-     * @param name the device identifier of the device calling the callback
-     * @param value the double data value
-     */
-    public void callback(String device, double value);
-
-}
