--- src/main/java/org/team199/wpiws/interfaces/ExceptionConsumer.java
@@ -1,11 +1,0 @@
-package org.team199.wpiws.interfaces;
-
-/**
- * Represents a function which is equivalent to {@link java.util.function.Consumer} with the ability to throw an {@link Exception}
- */
-@FunctionalInterface
-public interface ExceptionConsumer<T> {
-
-    public void accept(T t) throws Exception;
-
-}
