--- src/main/java/org/team199/wpiws/interfaces/package-info.java
@@ -1,4 +1,0 @@
-/**
- * All required {@link java.lang.FunctionalInterface} interfaces used in this API
- */
-package org.team199.wpiws.interfaces;
