--- src/main/java/org/team199/wpiws/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Code to interface with the WebSockets API of WPILib
- */
-package org.team199.wpiws;
